+++ test/common/upstream/ring_hash_lb_test.cc
@@ -247,6 +247,81 @@ TEST_P(RingHashLoadBalancerTest, BasicWithMurmur2) {
   EXPECT_EQ(0UL, stats_.lb_healthy_panic_.value());
 }
 
+// Expect reasonable results with hostname.
+TEST_P(RingHashLoadBalancerTest, BasicWithHostname) {
+  hostSet().hosts_ = {makeTestHost(info_, "90", "tcp://127.0.0.1:90"),
+                      makeTestHost(info_, "91", "tcp://127.0.0.1:91"),
+                      makeTestHost(info_, "92", "tcp://127.0.0.1:92"),
+                      makeTestHost(info_, "93", "tcp://127.0.0.1:93"),
+                      makeTestHost(info_, "94", "tcp://127.0.0.1:94"),
+                      makeTestHost(info_, "95", "tcp://127.0.0.1:95")};
+  hostSet().healthy_hosts_ = hostSet().hosts_;
+  hostSet().runCallbacks({}, {});
+
+  config_ = envoy::config::cluster::v3::Cluster::RingHashLbConfig();
+  config_.value().mutable_minimum_ring_size()->set_value(12);
+
+  common_config_ = envoy::config::cluster::v3::Cluster::CommonLbConfig();
+  auto chc = envoy::config::cluster::v3::Cluster::CommonLbConfig::ConsistentHashingLbConfig();
+  chc.set_use_hostname_for_hashing(true);
+  common_config_.set_allocated_consistent_hashing_lb_config(&chc);
+
+  init();
+  common_config_.release_consistent_hashing_lb_config();
+
+  EXPECT_EQ("ring_hash_lb.size", lb_->stats().size_.name());
+  EXPECT_EQ("ring_hash_lb.min_hashes_per_host", lb_->stats().min_hashes_per_host_.name());
+  EXPECT_EQ("ring_hash_lb.max_hashes_per_host", lb_->stats().max_hashes_per_host_.name());
+  EXPECT_EQ(12, lb_->stats().size_.value());
+  EXPECT_EQ(2, lb_->stats().min_hashes_per_host_.value());
+  EXPECT_EQ(2, lb_->stats().max_hashes_per_host_.value());
+
+  // hash ring:
+  // host | position
+  // ---------------------------
+  // 95 | 1975508444536362413
+  // 95 | 2376063919839173711
+  // 93 | 2386806903309390596
+  // 94 | 6749904478991551885
+  // 93 | 6803900775736438537
+  // 92 | 7225015537174310577
+  // 90 | 8787465352164086522
+  // 92 | 11282020843382717940
+  // 91 | 13723418369486627818
+  // 90 | 13776502110861797421
+  // 91 | 14338313586354474791
+  // 94 | 15364271037087512980
+
+  LoadBalancerPtr lb = lb_->factory()->create();
+  {
+    TestLoadBalancerContext context(0);
+    EXPECT_EQ(hostSet().hosts_[5], lb->chooseHost(&context));
+  }
+  {
+    TestLoadBalancerContext context(std::numeric_limits<uint64_t>::max());
+    EXPECT_EQ(hostSet().hosts_[5], lb->chooseHost(&context));
+  }
+  {
+    TestLoadBalancerContext context(7225015537174310577);
+    EXPECT_EQ(hostSet().hosts_[2], lb->chooseHost(&context));
+  }
+  {
+    TestLoadBalancerContext context(6803900775736438537);
+    EXPECT_EQ(hostSet().hosts_[3], lb->chooseHost(&context));
+  }
+  { EXPECT_EQ(hostSet().hosts_[5], lb->chooseHost(nullptr)); }
+  EXPECT_EQ(0UL, stats_.lb_healthy_panic_.value());
+
+  hostSet().healthy_hosts_.clear();
+  hostSet().runCallbacks({}, {});
+  lb = lb_->factory()->create();
+  {
+    TestLoadBalancerContext context(0);
+    EXPECT_EQ(hostSet().hosts_[5], lb->chooseHost(&context));
+  }
+  EXPECT_EQ(1UL, stats_.lb_healthy_panic_.value());
+}
+
 // Given 2 hosts and a minimum ring size of 3, expect 2 hashes per host and a ring size of 4.
 TEST_P(RingHashLoadBalancerTest, UnevenHosts) {
   hostSet().hosts_ = {makeTestHost(info_, "tcp://127.0.0.1:80"),